Here, let me serve your bitch lasagna

+++ src/pt/ulusofona/lp2/crazyChess/CrazyPiece.java
@@ -228,7 +228,7 @@ public abstract class CrazyPiece {
 
     public double getRacio() {
         if ((this.jogadasInvalidas + this.jogadasValidas) > 0) {
-            return (double) (this.jogadasInvalidas / (this.jogadasInvalidas + this.jogadasValidas));
+            return ((double)this.jogadasInvalidas / ((double)this.jogadasInvalidas + (double)this.jogadasValidas));
         }
         return 0;
     }
